Add Dathomir transport network screenplay for teleportation between Science Outpost and Aurillia Village
